requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dictionaries==0.0.2
distlib==0.3.6
Django==4.1.7
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.95.0
fastapi-users==10.4.1
fastapi-users-db-sqlalchemy==5.0.0
fastjsonschema==2.16.3
ffmpeg==1.4
filelock==3.10.7
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jose==1.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
limits==3.4.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openai==0.27.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.6
pure-eval==0.2.2
py-flags==1.1.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pyfmodex==0.7.0
pygame==2.1.3
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruff==0.0.254
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.13
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
